+++ R/cff_git_hook.R
@@ -8,7 +8,7 @@
 #'
 #' @name cff_git_hook
 #'
-#' @family Git helpers
+#' @family git
 #'
 #' @export
 #'