+++ R/docs.R
@@ -1,7 +1,7 @@
 #' The `cff` class
 #'
 #' @name cff-class
-#'
+#' @aliases cff_class
 #' @keywords internal
 #'
 #' @description