+++ man/roxygen/meta.R
@@ -1,0 +1,11 @@
+list(
+  rd_family_title = list(
+    core = "Other core functions of \\CRANpkg{cffr}:",
+    reading = "Reading other files as \\code{cff}:",
+    bibtex = "Other functions for working with BibTeX format:",
+    schemas = "Other CFF schemas:",
+    git = "Other Git/GitHub helpers:",
+    deprecated = "Other deprecated functions:",
+    s3method = "Other S3 Methods for \\code{cff}:"
+  )
+)