+++ man/roxygen2/meta.R
@@ -1,7 +1,0 @@
-list(
-  rd_family_title = list(
-    bibtex = "Coercing objects from/to BibTeX:",
-    git = "Git utils:",
-    schema = "CFF schemas:"
-  )
-)